[TC-40-refactor] - bugfix
