+++ src/main/java/com/server/booyoungee/domain/place/dao/movie/MoviePlaceRepository.java
@@ -2,14 +2,12 @@ package com.server.booyoungee.domain.place.dao.movie;
 
 import java.util.List;
 
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import com.server.booyoungee.domain.place.domain.moviePlace.MoviePlace;
-import com.server.booyoungee.domain.place.domain.storePlace.StorePlace;
+import com.server.booyoungee.domain.place.domain.movie.MoviePlace;
 
 @Repository
 public interface MoviePlaceRepository extends JpaRepository<MoviePlace, Long> {
